Move proxy contract out of subsidies
