+++ gossip/blockproc/proxy/proxy.sol
