+++ gossip/blockproc/proxy/proxy_abigen.go
