+++ gossip/blockproc/proxy/proxy_contract.bin
