+++ integration/makefakegenesis/json.go
@@ -25,7 +25,7 @@ import (
 	"os"
 	"time"
 
-	"github.com/0xsoniclabs/sonic/gossip/blockproc/subsidies/proxy"
+	"github.com/0xsoniclabs/sonic/gossip/blockproc/proxy"
 	"github.com/0xsoniclabs/sonic/gossip/blockproc/subsidies/registry"
 	"github.com/0xsoniclabs/sonic/integration/makegenesis"
 	"github.com/0xsoniclabs/sonic/inter"
